+++ packages/tiptap-extensions/web/src/placeholder.ts
@@ -5,6 +5,7 @@ import type {
   RicosExtension,
   RicosExtensionConfig,
   RicosNodeExtension,
+  RicosServices,
 } from 'ricos-types';
 import { compact } from 'lodash';
 
@@ -15,7 +16,9 @@ export const placeholder: RicosExtension = {
   reconfigure: (
     config: RicosExtensionConfig,
     _extensions: RicosExtension[],
-    ricosProps: ExtensionProps
+    ricosProps: ExtensionProps,
+    _settings: Record<string, unknown>,
+    services: RicosServices
   ) => ({
     ...config,
     addOptions() {
@@ -26,6 +29,7 @@ export const placeholder: RicosExtension = {
         showOnlyWhenEditable: true,
         showOnlyCurrent: true,
         includeChildren: false,
+        t: services.t,
         extensionsPlaceholders: compact(
           _extensions
             .filter(ext => ext.type === 'node')
@@ -92,17 +96,20 @@ export const placeholder: RicosExtension = {
                   const isEmpty = !node.isLeaf && !node.childCount;
                   if (isEmpty) {
                     if (this.options.extensionsPlaceholders.length) {
-                      this.options.extensionsPlaceholders.forEach(({ predicate, content }) => {
-                        if (predicate({ doc, pos, node })) {
-                          const classes = [this.options.emptyNodeClass];
-                          const decoration = Decoration.node(pos, pos + node.nodeSize, {
-                            class: classes.join(' '),
-                            'data-placeholder': content,
-                          });
+                      this.options.extensionsPlaceholders.forEach(
+                        ({ predicate, translationKey, content }) => {
+                          if (predicate({ doc, pos, node })) {
+                            const { t } = this.options;
+                            const classes = [this.options.emptyNodeClass];
+                            const decoration = Decoration.node(pos, pos + node.nodeSize, {
+                              class: classes.join(' '),
+                              'data-placeholder': translationKey && t ? t(translationKey) : content,
+                            });
 
-                          decorations.push(decoration);
+                            decorations.push(decoration);
+                          }
                         }
-                      });
+                      );
 
                       return true;
                     }